+++ extensions/sn_debug_info/change_log.md
@@ -5,3 +5,5 @@ Change Log
   - Start of log.
 * 1.1
   - Added totals to station ware users.
+* 1.2
+  - Update for x4 7.0.